port LoadResourceAsync; widen ClearBit/WriteBit ptr cast; gUnk_081092AC zero-buffer override
Agent-Logs-Url: https://github.com/awe444/tmc/sessions/08febf1f-f8e0-4506-a43d-02a4289cb3f6

+++ src/common.c
@@ -122,12 +122,12 @@ u32 ReadBit(void* src, u32 bit) {
 }
 
 u32 WriteBit(void* src, u32 bit) {
-    u32 b;
+    uintptr_t b;
     u32 mask;
     u32 orig;
 
     // note that the following line relies on undefined behaviour; b is not initialised
-    b += ((u32)src + bit / 8) - b;
+    b += ((uintptr_t)src + bit / 8) - b;
     mask = 0x7;
     mask = 1 << (bit & mask);
 
@@ -139,12 +139,12 @@ u32 WriteBit(void* src, u32 bit) {
 }
 
 u32 ClearBit(void* src, u32 bit) {
-    u32 b;
+    uintptr_t b;
     u32 mask;
     u32 orig;
 
     // note that the following line relies on undefined behaviour; b is not initialised
-    b += ((u32)src + bit / 8) - b;
+    b += ((uintptr_t)src + bit / 8) - b;
     mask = 0x7;
     mask = 1 << (bit & mask);
